new logic to detect unconfigured firewalls. enable mdns in this case.
This shall help in cases like bnc#959707.

I guess many users don't think of configuring their firewall. In the
desktop case most users don't have different zones any more. Firewalling
is done by a hardware router. They want to enjoy zeroconf services. mdns
is the basis for this. So we're allowing it in case the firewall zones
are unconfigured.