conntrack helpers: support kernels >= 4.7 with nf_conntrack_helper = 0
bnc#986527: newer kernels disabled implicit use of conntrack helpers. We
now need to setup them explicitly which is somewhat complicated for the
different connection tracker helpers out there. I've implemented this
with support for all conntrack helpers in mind, but it will probably
only work for the netbios-ns case at the moment, because the others need
slightly different rules in the raw table.