FW_ROUTE: harmonize setting of ip forwarding for IPv4/6 (bnc#572202)
yast2 network and SuSEfirewall shall no longer depend on each other. For
backward compatibility SuSEfirewall2 can still set ip forwarding while
it's enabled. It will do nothing any more if it is already enabled,
however.

Also it will restore ip forwarding to the original values from
/etc/sysctl.conf after stopping it, if applicable.

It will no longer reset forwarding to 0 if FW_ROUTE is not set.

Also this commit brings a clearer separation of routing for IPv4 and
IPv6, allowing to reduce the amount of forwarding rules if either of
both is not used at all.