systemd: fix a race condition in the _init service (bsc#1014987)
As suggested in the bug report I've added further requirements to the
systemd unit for the init service. Also I've switched the mktemp call to
use /dev/shm instead of /tmp which should allow the service to be
executed earlier and doesn't rely on an r/w /tmp.

cherry pick from master.