removal of SuSEfirewall2 from factory: add services and fail2ban config
This adds the previously decentrally managed SuSEfirewall2 service files
to the SuSEfirewall2 repository. For people that for some reason still
want to use SuSEfirewall2 this is a simple way to get their hands on
them without looking in a plethora of OBS packages.

The SuSEfirewall2-fail2ban sub-package is also going to be removed from
openSUSE:Factory, therefore also add this bit of configuration to the
main repository.